+++ backend/src/neuroagent/app/dependencies.py
@@ -54,6 +54,7 @@ from neuroagent.tools import (
     ExperimentalNeuronDensityGetOneTool,
     ExperimentalSynapsesPerConnectionGetAllTool,
     ExperimentalSynapsesPerConnectionGetOneTool,
+    GenerateSimulationsConfigTool,
     IonChannelModelGetAllTool,
     IonChannelModelGetOneTool,
     LinkGenerationTool,
@@ -403,6 +404,7 @@ def get_tool_list(
         SingleNeuronSynaptomeSimulationGetOneTool,
         CircuitGetAllTool,
         CircuitGetOneTool,
+        GenerateSimulationsConfigTool,
         # NowTool,
         # WeatherTool,
         # RandomPlotGeneratorTool,
@@ -628,6 +630,7 @@ async def get_context_variables(
         "s3_client": s3_client,
         "thread_id": thread.thread_id,
         "thumbnail_generation_url": settings.tools.thumbnail_generation.url,
+        "usage_dict": {},
         "user_id": user_info.sub,
         "vlab_id": thread.vlab_id,
     }